refactor: Move PrusaLink and Moonraker models to shared project (Phase 1)
- Create src/shared/Contracts/Printers/PrusaLinkModels.cs with namespace Farm.Web.Shared.Contracts.Printers.PrusaLink
- Create src/shared/Contracts/Printers/MoonrakerModels.cs with namespace Farm.Web.Shared.Contracts.Printers.Moonraker
- Delete original api/Services/PrusaLinkModels.cs and api/Services/MoonrakerModels.cs
- Rename DirectoryInfo -> MoonrakerDirectoryInfo to avoid System.IO conflict
- Rename FileInfo -> PrusaLinkFileInfo to avoid System.IO conflict
- Remove duplicate History types from MoonrakerModels (use Farm.Web.Shared versions)
- Update all API and test file imports to use new namespaces

Build: 0 Warnings, 0 Errors

+++ src/api/Mapping/PrinterMappingProfile.cs
@@ -1,5 +1,7 @@
 ﻿using AutoMapper;
 using Farm.Web.Api.Services;
+using Farm.Web.Shared;
+using Farm.Web.Shared.Contracts.Printers.Moonraker;
 using SharedModels = Farm.Web.Shared;
 
 namespace Farm.Web.Api.Mapping
@@ -30,7 +32,7 @@ namespace Farm.Web.Api.Mapping
                 .ForMember(dest => dest.AuxiliaryTotals, opt => opt.MapFrom(src => src.AuxiliaryTotals));
 
             _ = CreateMap<JobTotals, SharedModels.JobTotals>()
-                .ForMember(dest => dest.TotalJobs, opt => opt.MapFrom(src => (int)src.TotalJobs))
+                .ForMember(dest => dest.TotalJobs, opt => opt.MapFrom(src => src.TotalJobs))
                 .ForMember(dest => dest.TotalTime, opt => opt.MapFrom(src => src.TotalTime))
                 .ForMember(dest => dest.TotalPrintTime, opt => opt.MapFrom(src => src.TotalPrintTime))
                 .ForMember(dest => dest.TotalFilamentUsed, opt => opt.MapFrom(src => src.TotalFilamentUsed))
